Added buttons, refactored picturelamp to store values in scaled integer

+++ Arduino/ChistmasLights/picturelamp.h
@@ -23,8 +23,8 @@ class PictureLamp
   protected:
       unsigned long start_transition_time=0;
       unsigned long transition_duration=0; /* 0 = not in transition */
-      float current_red,current_green,current_blue;
-      float target_red,target_green,target_blue;
+      int current_red,current_green,current_blue;
+      int target_red,target_green,target_blue;
       transition_type_t transition_type;
 };
